small refactoring to the AnswerList, Move the current rooms answers statistics into the component, add readOnly variant for the outro screen, add it the outro state

+++ src/jsMain/kotlin/components/AnswerList.kt
@@ -1,7 +1,9 @@
 package components
 
 import api.RoomApi
+import com.fynnian.application.common.I18n
 import com.fynnian.application.common.room.Answer
+import com.fynnian.application.common.room.RoomImage
 import csstype.*
 import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.launch
@@ -15,6 +17,7 @@ import mui.system.sx
 import react.FC
 import react.Props
 import react.dom.onChange
+import react.useContext
 import react.useState
 import web.html.HTMLTextAreaElement
 import web.html.InputType
@@ -22,19 +25,46 @@ import web.html.InputType
 private val scope = MainScope()
 
 external interface AnswerListProps : Props {
+  var images: List<RoomImage>
   var answers: List<Answer>
   var reloadAnswers: () -> Unit
+  var readOnly: Boolean?
+  var totalAnswers: Int?
 }
 
 val AnswerList = FC<AnswerListProps> { props ->
-  List {
-    sx {
+  val (language) = useContext(LanguageContext)
 
+  Box {
+    sx {
+      padding = 0.5.rem
     }
+    Typography {
+      variant = TypographyVariant.body1
+      +I18n.get(
+        language,
+        I18n.TranslationKey.ROOM_ANSWER_ANSWERS_TOTAL,
+        I18n.TemplateProperty.Answers(props.totalAnswers ?: props.answers.size)
+      )
+    }
+    if (props.images.size > 1 && props.readOnly != true) props.images.mapIndexed { i, image ->
+      Typography {
+        variant = TypographyVariant.body1
+        +I18n.get(
+          language,
+          I18n.TranslationKey.ROOM_ANSWER_ANSWERS_COUNT_PER_IMAGE,
+          I18n.TemplateProperty.Number(i + 1),
+          I18n.TemplateProperty.Answers(props.answers.filter { it.imageId == image.id }.size)
+        )
+      }
+    }
+  }
+  List {
     props.answers.map {
       AnswerListItem {
         answer = it
         reloadAnswers = props.reloadAnswers
+        readOnly = props.readOnly ?: false
       }
     }
   }
@@ -43,6 +73,7 @@ val AnswerList = FC<AnswerListProps> { props ->
 external interface AnswerListItemProps : Props {
   var answer: Answer
   var reloadAnswers: () -> Unit
+  var readOnly: Boolean
 }
 
 val AnswerListItem = FC<AnswerListItemProps> { props ->
@@ -115,20 +146,22 @@ val AnswerListItem = FC<AnswerListItemProps> { props ->
         }
         +answer.answer
       }
-      IconButton {
-        color = IconButtonColor.primary
-        onClick = { edit = true }
-        Edit()
-      }
-      IconButton {
-        color = IconButtonColor.primary
-        onClick = {
-          scope.launch {
-            api.deleteAnswer(answer.roomCode, answer.id)
-            props.reloadAnswers()
-          }
+      if (!props.readOnly){
+        IconButton {
+          color = IconButtonColor.primary
+          onClick = { edit = true }
+          Edit()
         }
-        Delete()
+        IconButton {
+          color = IconButtonColor.primary
+          onClick = {
+            scope.launch {
+              api.deleteAnswer(answer.roomCode, answer.id)
+              props.reloadAnswers()
+            }
+          }
+          Delete()
+        }
       }
     }
   }
